- oracle zfs handling improved

+++ lib/ZnapZend/ZFS.pm
@@ -9,7 +9,7 @@ use POSIX qw(WNOHANG SIGTERM SIGKILL);
 has debug           => sub { 0 };
 has noaction        => sub { 0 };
 has nodestroy       => sub { 1 };
-has combinedDestroy => sub { 0 };
+has oracleMode      => sub { 0 };
 has recvu           => sub { 0 };
 has sendDelay       => sub { 3 };
 has propertyPrefix  => sub { q{org.znapzend} };
@@ -210,36 +210,41 @@ sub destroySnapshots {
     my %toDestroy;
     my ($remote, $dataSet, $snapshot);
 
-    #combinedDestroy flag set...
-    if ($self->combinedDestroy){
+    #oracleMode: destroy each snapshot individually
+    if ($self->oracleMode){
+        my $destroyError = '';
         for (@toDestroy){
             ($remote, $dataSet) = $splitHostDataSet->($_);
-            ($dataSet, $snapshot) = $splitDataSetSnapshot->($dataSet);
-            #tag local snapshots as 'local' so we have a key to build the hash
-            $remote = $remote || 'local';
-            exists $toDestroy{$remote} or $toDestroy{$remote} = [];
-            push @{$toDestroy{$remote}}, @{$toDestroy{$remote}} ? $snapshot : "$dataSet\@$snapshot";
-        }
-
-        for $remote (keys %toDestroy){
-            #check if remote is flaged as 'local'.
-            my @ssh = $self->$buildRemote($remote ne 'local'
-                ? $remote : undef, [qw(zfs destroy), join(',', @{$toDestroy{$remote}})]);
+            my @ssh = $self->$buildRemote($remote, [qw(zfs destroy), $dataSet]);
 
             print STDERR '# ' . join(' ', @ssh) . "\n" if $self->debug;
-            system(@ssh) && Mojo::Exception->throw("ERROR: cannot destroy snapshot(s) $toDestroy[0]")
+            system(@ssh) and $destroyError .= "ERROR: cannot destroy snapshot $dataSet\n"
                 if !($self->noaction || $self->nodestroy);
         }
+        #remove trailing \n
+        chomp $destroyError;
+        Mojo::Exception->throw($destroyError) if $destroyError ne '';
+
         return 1;
     }
 
-    #destroy each snapshot individually
+    #combinedDestroy
     for (@toDestroy){
         ($remote, $dataSet) = $splitHostDataSet->($_);
-        my @ssh = $self->$buildRemote($remote, [qw(zfs destroy), $dataSet]);
+        ($dataSet, $snapshot) = $splitDataSetSnapshot->($dataSet);
+        #tag local snapshots as 'local' so we have a key to build the hash
+        $remote = $remote || 'local';
+        exists $toDestroy{$remote} or $toDestroy{$remote} = [];
+        push @{$toDestroy{$remote}}, @{$toDestroy{$remote}} ? $snapshot : "$dataSet\@$snapshot";
+    }
+
+    for $remote (keys %toDestroy){
+        #check if remote is flaged as 'local'.
+        my @ssh = $self->$buildRemote($remote ne 'local'
+            ? $remote : undef, [qw(zfs destroy), join(',', @{$toDestroy{$remote}})]);
 
         print STDERR '# ' . join(' ', @ssh) . "\n" if $self->debug;
-        system(@ssh) && Mojo::Exception->throw("ERROR: cannot destroy snapshot $dataSet")
+        system(@ssh) && Mojo::Exception->throw("ERROR: cannot destroy snapshot(s) $toDestroy[0]")
             if !($self->noaction || $self->nodestroy);
     }
